pyproject.toml
[project]
name = "bp-pred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "h5py>=3.15.1",
    "jupyter>=1.1.1",
    "mat73>=0.62",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "scipy>=1.17.1",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.2.0",
]
